fix: run pipeline modal stage selection (#7827)
